Functional version, spatial information added, CNN fixed
Spatial information (distance between the two persons) is passed to the logical part through a pre-constructed .pl file consisting of facts of the form distance(video, p1, p2, timestep, D) for each timestep, where D denotes the distance between the people. This file is consulted within caviar.pl.
Fixed the CNN forward() function in order to access the correct part of the input. Input is (295, 24, 2, 3, 80, 80) so get the correct (3, 80, 80) image by using the personID, and timestep arguments.